Update startup-date.sh helper

+++ helpers/startup-date.sh
@@ -1,11 +1,13 @@
 #!/bin/bash
 
 setStartupDate() {
-    STARTUP_DATE=$( date +"%s" )
-    touch "$ENV_FILE_PATH"
-    echo "$STARTUP_DATE" > "$ENV_FILE_PATH"
+  STARTUP_DATE=$( date +"%s" )
+  touch "$ENV_FILE_PATH"
+  echo "$STARTUP_DATE" > "$ENV_FILE_PATH"
 }
 
 resetStartupDate() {
-    rm "$ENV_FILE_PATH"
+  if test -f "$ENV_FILE_PATH"; then
+      rm "$ENV_FILE_PATH"
+  fi
 }
